updated advisor flow and added it to budgetItems and tasks

+++ suruga_home/Price.h
@@ -21,6 +21,6 @@
 
 - (double) getSum;
 - (double) getInitialSum;
-- (double) getRunningSum;
+- (int) getRunningSum;
 
 @end
